Moved V2 versions of class files to V2 specific folder.
